+++ assets/src/anycomment/src/assets/js/app.js
@@ -1,7 +1,8 @@
 import $ from 'jquery';
 // import whatInput from 'what-input';
-import 'select2'
-import '@claviska/jquery-minicolors'
+import 'select2';
+import '@claviska/jquery-minicolors';
+import introJs from 'intro.js';
 
 import CodeMirror from 'codemirror';
 import 'codemirror/mode/css/css';
@@ -14,6 +15,45 @@ window.$ = $;
 // If you want to pick and choose which modules to include, comment out the above and uncomment
 // the line below
 //import './lib/foundation-explicit-pieces';
+
+/**
+ * Starts tour on admin functionality.
+ *
+ * When tour should be enable, this is not executed internally.
+ */
+function maybeStartTour() {
+    const intro = introJs();
+    const AnyComment = window.AnyComment || {};
+    const tourSteps = AnyComment.TourSteps || {};
+
+    const stepsPageMap = tourSteps.steps || {};
+
+    const wrapperRoot = document.getElementById('anycomment-wrapper');
+
+    if (wrapperRoot && wrapperRoot.dataset) {
+        const introShow = parseInt(wrapperRoot.dataset.introShow);
+        const currentTab = wrapperRoot.dataset.currentTab || '';
+
+        if (introShow === 1) {
+            if (stepsPageMap[currentTab]) {
+                intro.setOptions({
+                    steps: stepsPageMap[currentTab],
+                    nextLabel: 'Далее',
+                    prevLabel: 'Назад',
+                    skipLabel: 'Пропустить',
+                    doneLabel: 'Скрыть',
+                });
+                intro.start();
+
+                intro.onexit(() => {
+                    if (typeof tourSteps.onExit === 'function') {
+                        tourSteps.onExit('onexit', intro);
+                    }
+                });
+            }
+        }
+    }
+}
 
 // jQuery(document).foundation();
 
@@ -46,4 +86,6 @@ $(document).ready(function ($) {
             });
         }
     }
+
+    maybeStartTour();
 });