+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Freemius was cut-out. This removes problems for some websites + speeds-up plugin a bit.
 - Added support for BuddyPress, when somebody send "@" + username in comment. This user would be mentioned, #422
 - Moved AnyComment Cloud message to Integration tab, improved UI/UX of synchronization logic 
+- Added helper tour around plugin to help new users to know how to configure plugin
 
 **Fixes:**